add setTelemetry
