+++ lib/index.js
@@ -1,8 +1,10 @@
 const gatherTelemetryForUrl = require('./gather/gather-telemetry');
-const { getTelemetryFor, getModulePathFor, getTelemetry } = require('./utils/get-telemetry-for');
+const { getTelemetryFor, getModulePathFor } = require('./utils/get-telemetry-for');
+const { getTelemetry, setTelemetry } = require('./utils/telemetry');
 
 module.exports = {
   getTelemetry,
+  setTelemetry,
   gatherTelemetryForUrl,
   getTelemetryFor,
   getModulePathFor,
